[#233] Relocate JEOD .cc parser out of production jeod_gravity (#232 Wave 0)
Move load_from_jeod_cc, load_mu_from_jeod_cc, CoeffLoadError, and the
private parsing helpers from jeod_gravity::coefficients into a new
jeod_test_data::jeod_cc module. Production gravity no longer knows
how to parse JEOD source files; only the dev/test crate does.

Foundation step for #232 (phasing JEOD_HOME out of test runtime).
Production types (SphericalHarmonicsData, save_binary, load_binary,
load_binary_from_bytes) are unchanged.

Pre-creates pub mod stubs for gravity_fixtures, body_init_fixtures,
mars_fixtures so Wave 1 worktrees (#234/#235/#236) don't conflict on
jeod_test_data/src/lib.rs while developing in parallel.

Callers updated across jeod_gravity tests, jeod_runner tests and
verification rigs, jeod_sim verification recipes, and the bevy_parity
high-fidelity integration test.
