+++ crates/jeod_runner/src/run_verification/sim_solar_beta.rs
@@ -30,10 +30,11 @@ use jeod_sim::recipes::verification::{
     CsvReference, ExtrasComparator, InitialConditions, PreStepClosure, Tolerances, VerificationCase,
 };
 use jeod_sim::{
-    coefficients, default_leap_second_table, DerivedStateConfig, Ephemeris, EphemerisBody,
-    GravityControl, GravityControls, GravityModel, GravitySource, GravitySourceEntry,
-    RotationModel, SimulationBuilder, SimulationTime, TranslationalState, VehicleConfig,
+    default_leap_second_table, DerivedStateConfig, Ephemeris, EphemerisBody, GravityControl,
+    GravityControls, GravityModel, GravitySource, GravitySourceEntry, RotationModel,
+    SimulationBuilder, SimulationTime, TranslationalState, VehicleConfig,
 };
+use jeod_test_data::jeod_cc as coefficients;
 use uom::si::f64::Time;
 use uom::si::time::second;
