+++ crates/jeod_runner/tests/tier3_sim_orbinit_edge.rs
@@ -29,7 +29,7 @@ fn tier3_simulation_orbinit_cross_consistency() {
 
     let grav_data_dir = jeod_root.join("models/environment/gravity/data/src");
     let mu_earth =
-        jeod_sim::coefficients::load_mu_from_jeod_cc(&grav_data_dir.join("earth_GGM05C.cc"))
+        jeod_test_data::jeod_cc::load_mu_from_jeod_cc(&grav_data_dir.join("earth_GGM05C.cc"))
             .expect("load Earth mu");
 
     let runs: Vec<(&str, &str)> = vec![
