+++ crates/jeod_runner/tests/tier3_sim_time_reversal.rs
@@ -22,7 +22,7 @@ fn load_mu_earth_gemt1() -> f64 {
         "JEOD source not found at {}. Set JEOD_HOME or JEOD_PATH.",
         jeod_root.display()
     );
-    jeod_sim::coefficients::load_mu_from_jeod_cc(
+    jeod_test_data::jeod_cc::load_mu_from_jeod_cc(
         &jeod_root.join("models/environment/gravity/data/src/earth_GEMT1.cc"),
     )
     .expect("load Earth mu from GEMT1")
