+++ crates/jeod_sim/Cargo.toml
@@ -8,6 +8,16 @@ edition.workspace = true
 license.workspace = true
 repository.workspace = true
 
+[features]
+# `jeod-source` exposes verification helpers that read from a JEOD source
+# checkout (`$JEOD_HOME` / `$JEOD_PATH`) — currently the
+# `recipes::verification::reference_data` loaders that parse `.cc`
+# gravity files via `jeod_test_data::jeod_cc`. Off by default so
+# downstream mission crates don't pull `jeod_test_data` (a dev-flavored
+# crate) into their production build. Verification rigs in
+# `jeod_runner` enable this feature.
+jeod-source = ["dep:jeod_test_data"]
+
 [dependencies]
 jeod_quantities = { path = "../jeod_quantities", version = "0.1.0" }
 jeod_math = { path = "../jeod_math", version = "0.1.0" }
@@ -19,6 +29,11 @@ jeod_atmosphere = { path = "../jeod_atmosphere", version = "0.1.0" }
 jeod_interactions = { path = "../jeod_interactions", version = "0.1.0" }
 jeod_ephemeris = { path = "../jeod_ephemeris", version = "0.1.0" }
 jeod_planet = { path = "../jeod_planet", version = "0.1.0" }
+# Optional: only pulled in when the `jeod-source` feature is enabled.
+# Used by `recipes::verification::reference_data` for parsing JEOD `.cc`
+# coefficient files. `jeod_test_data` is the dev/test-data crate; this
+# is the single non-dev consumer.
+jeod_test_data = { path = "../jeod_test_data", version = "0.1.0", optional = true }
 glam = { workspace = true }
 log = { workspace = true }
 uom = { workspace = true }
